Fix React Suspense (#12)
* fix collapsing headers

* fix suspends
