+++ src/pages/ResourceCategoryDetail/index.tsx
@@ -1,25 +1,9 @@
-import {
-  IonBackButton,
-  IonButtons,
-  IonContent,
-  IonGrid,
-  IonHeader,
-  IonPage,
-  IonRow,
-  IonTitle,
-  IonToolbar,
-  IonCol,
-  IonItem,
-  IonLabel,
-  IonList,
-  IonItemDivider,
-} from '@ionic/react';
-import Interweave from 'interweave';
-import React from 'react';
+import { IonPage } from '@ionic/react';
+import React, { Suspense } from 'react';
 import { RouteComponentProps } from 'react-router';
-import { useResource } from 'rest-hooks';
-import ResourceCategoryResource from '../../resources/resourceCategory';
-import ResourcePageResource from '../../resources/resourcePage';
+import { NetworkErrorBoundary } from 'rest-hooks';
+import SuspenseFallback from '../../components/SuspenseFallback';
+import ResourceCategoryDetailInner from './ResourceCategoryDetailInner';
 
 type ResourceCategoryDetailProps = RouteComponentProps<{
   id: string;
@@ -30,47 +14,13 @@ const ResourceCategoryDetail: React.FC<ResourceCategoryDetailProps> = ({
     params: { id },
   },
 }) => {
-  const [category, pages] = useResource(
-    [ResourceCategoryResource.detailShape(), { id }],
-    [ResourcePageResource.listShape(), { 'filter[category][=]': id }]
-  );
-
   return (
     <IonPage>
-      <IonHeader collapse="condense">
-        <IonToolbar>
-          <IonButtons slot="start">
-            <IonBackButton defaultHref="/resources" text="Resources" />
-          </IonButtons>
-        </IonToolbar>
-        <IonToolbar>
-          <IonTitle size="large">{category.name}</IonTitle>
-        </IonToolbar>
-      </IonHeader>
-      <IonContent>
-        <IonGrid>
-          <IonRow>
-            <IonCol>
-              <Interweave content={category.description}></Interweave>
-            </IonCol>
-          </IonRow>
-        </IonGrid>
-
-        <IonList>
-          <IonItemDivider>
-            <IonLabel>Pages</IonLabel>
-          </IonItemDivider>
-          {pages.map((page) => (
-            <IonItem
-              key={page.id}
-              lines="full"
-              routerLink={`/pages/${page.id}`}
-            >
-              <IonLabel>{page.title}</IonLabel>
-            </IonItem>
-          ))}
-        </IonList>
-      </IonContent>
+      <NetworkErrorBoundary>
+        <Suspense fallback={<SuspenseFallback />}>
+          <ResourceCategoryDetailInner id={id} />
+        </Suspense>
+      </NetworkErrorBoundary>
     </IonPage>
   );
 };
