+++ src/pages/ResourceCategoryList/index.tsx
@@ -1,51 +1,18 @@
-import {
-  IonCol,
-  IonContent,
-  IonGrid,
-  IonHeader,
-  IonPage,
-  IonRow,
-  IonTitle,
-  IonToolbar,
-} from '@ionic/react';
-import React from 'react';
-import { useResource, useRetrieve } from 'rest-hooks';
-import FileResource from '../../resources/file';
-import ResourceCategoryResource from '../../resources/resourceCategory';
-import ResourceCategoryTile from './ResourceCategoryTile';
-
+import { IonPage } from '@ionic/react';
+import React, { Suspense } from 'react';
+import { NetworkErrorBoundary } from 'rest-hooks';
+import SuspenseFallback from '../../components/SuspenseFallback';
+import ResourceCategoryListInner from './ResourceCategoryListInner';
 import './resources.css';
 
 const ResourceCategoryList: React.FC = () => {
-  const resourceCategories = useResource(
-    ResourceCategoryResource.listShape(),
-    {}
-  );
-
-  // pre-cache icons
-  useRetrieve(FileResource.listShape(), {
-    'filter[id][in]': resourceCategories.map((cat) => cat.icon).join(','),
-  });
-
   return (
     <IonPage>
-      <IonHeader>
-        <IonToolbar></IonToolbar>
-        <IonToolbar>
-          <IonTitle size="large">Resources</IonTitle>
-        </IonToolbar>
-      </IonHeader>
-      <IonContent>
-        <IonGrid>
-          <IonRow>
-            {resourceCategories.map((category) => (
-              <IonCol key={category.id} size="6" class="resource-column">
-                <ResourceCategoryTile category={category} />
-              </IonCol>
-            ))}
-          </IonRow>
-        </IonGrid>
-      </IonContent>
+      <NetworkErrorBoundary>
+        <Suspense fallback={<SuspenseFallback title="Resources" />}>
+          <ResourceCategoryListInner></ResourceCategoryListInner>
+        </Suspense>
+      </NetworkErrorBoundary>
     </IonPage>
   );
 };
