+++ src/pages/Rooms/index.tsx
@@ -20,6 +20,11 @@ const Rooms: React.FC = () => {
         </IonToolbar>
       </IonHeader>
       <IonContent>
+        <IonHeader collapse="condense">
+          <IonToolbar>
+            <IonTitle size="large">Getting Around</IonTitle>
+          </IonToolbar>
+        </IonHeader>
         <IonGrid>
           <IonRow>
             <IonCol>
